Modularize FastAPI app structure:
- Extract `lifespan`, middleware, routers, and exception handlers into separate modules under `app`.
- Simplify `main.py` by delegating responsibilities to newly created modules.
- Refactor lifecycle to use `get_rabbit_broker` and improve shutdown handling.
- Enhance exception handling with a base `AppError` class for custom application errors.
